feat(wallet-sdk): own the app QueryClient (Phase 1 — the cache seam)
The SDK now owns the single TanStack QueryClient the web app uses — the riskiest seam of
the extraction — proven behavior-neutral before any domain logic moves.

- @agicash/wallet-sdk gains getQueryClient() (packages/wallet-sdk/src/query-client.ts): the
  exact isServer/browser-singleton pattern from the web, now on @tanstack/query-core (pinned
  to 5.90.20 = the version @tanstack/react-query depends on, incl. the workspace patch — so
  this QueryClient is the SAME class react-query's hooks use and the web can mount it directly)
- web's features/shared/query-client.ts is now a thin re-export of the SDK's getQueryClient,
  so all ~16 call sites (root provider, the supabase-session module-load capture, loaders,
  auth/user/cashu) are untouched and the browser singleton has a single home in the SDK
- web now declares @agicash/wallet-sdk (workspace:*); also declares @agicash/cashu, which the
  Phase-0 cashu shims import but the app had never declared (phantom-dep fix)
- SDK test locks the per-request (server) vs singleton (browser) contract

typecheck, biome, build all green; 135 tests (web 84 + utils 14 + cashu 35 + sdk 2); zero web
behavior change; @agicash/wallet-sdk is framework-free (query-core only).
