+++ apps/web-wallet/app/features/shared/query-client.ts
@@ -1,23 +1,3 @@
-import { QueryClient, isServer } from '@tanstack/react-query';
-
-let browserQueryClient: QueryClient | undefined = undefined;
-
-function makeQueryClient() {
-  return new QueryClient();
-}
-
-export function getQueryClient() {
-  if (isServer) {
-    // Server: always make a new query client
-    return makeQueryClient();
-  }
-
-  // Browser: make a new query client if we don't already have one
-  // This is very important, so we don't re-make a new client if React
-  // suspends during the initial render. This may not be needed if we
-  // have a suspense boundary BELOW the creation of the query client
-  if (!browserQueryClient) {
-    browserQueryClient = makeQueryClient();
-  }
-  return browserQueryClient;
-}
+// The QueryClient is owned by @agicash/wallet-sdk and re-exported here so the app's
+// getQueryClient() call sites stay unchanged. Removed in the import-cleanup PR.
+export { getQueryClient } from '@agicash/wallet-sdk/query-client';
