+++ apps/web-wallet/package.json
@@ -19,10 +19,12 @@
   "dependencies": {
     "@agicash/bc-ur": "0.1.0",
     "@agicash/breez-sdk-spark": "0.13.5-1",
+    "@agicash/cashu": "workspace:*",
     "@agicash/db-types": "workspace:*",
     "@agicash/opensecret": "catalog:",
     "@agicash/qr-scanner": "0.1.2",
     "@agicash/utils": "workspace:*",
+    "@agicash/wallet-sdk": "workspace:*",
     "@cashu/cashu-ts": "catalog:",
     "@noble/ciphers": "1.3.0",
     "@noble/curves": "catalog:",
