+++ packages/wallet-sdk/src/index.ts
@@ -1,1 +1,1 @@
-// @agicash/wallet-sdk
+export * from './query-client';
