+++ packages/wallet-sdk/tsconfig.json
@@ -3,7 +3,8 @@
   "include": ["src/**/*.ts"],
   "exclude": ["node_modules"],
   "compilerOptions": {
-    "lib": ["ES2022"],
+    "lib": ["ES2022", "DOM"],
+    "types": ["bun"],
     "noEmit": true
   }
 }
